Builds on msvc but controls are messed up. Video mode also seems to be applying android video mode to windows

+++ BattleNetwork/bnAudioResourceManager.h
@@ -82,8 +82,8 @@ private:
   Channel* channels;
   sf::SoundBuffer* sources;
   sf::Music stream;
-  int channelVolume;
-  int streamVolume;
+  float channelVolume;
+  float streamVolume;
   bool isEnabled;
 };
